+++ tests/test_delete.py
@@ -4,6 +4,7 @@ from scim2_models import Error
 from scim2_models import Group
 from scim2_models import User
 
+from scim2_client import RequestNetworkError
 from scim2_client import SCIMClient
 from scim2_client import SCIMRequestError
 
@@ -81,3 +82,14 @@ def test_dont_check_response_payload(httpserver):
         "status": "404",
         "detail": "404 error",
     }
+
+
+def test_request_network_error(httpserver):
+    """Test that httpx exceptions are transformed in RequestNetworkError."""
+
+    client = Client(base_url=f"http://localhost:{httpserver.port}")
+    scim_client = SCIMClient(client, resource_types=(User,))
+    with pytest.raises(
+        RequestNetworkError, match="Network error happened during request"
+    ):
+        scim_client.delete(User, "anything", url="http://invalid.test")